+++ src/alda/AldaRequestOptions.java
@@ -1,7 +1,6 @@
 package alda;
 
 public class AldaRequestOptions {
-  public String as;
   public String filename;
   public String from;
   public String to;